Added train stops and update them

+++ src/main/java/com/trs/trs_admin_service/repo/TrainRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TrainRepository extends JpaRepository<Train, Long> {
+public interface TrainRepository extends JpaRepository<Train, Integer> {
     Train findByTrainNumber(Integer trainNumber);
 }
